% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Graphical user interface for sample size estimation with mixed outcome composite endpoints}
\usage{
gui()
}
\description{
\code{gui()} runs an R Shiny web browser based graphical user interface for
\code{\link{multsampsize}}. For details of Shiny app functionality, 
see \href{https://github.com/martinamcm/MultSampSize}{MultSampSize}
}
\examples{
# Launch the graphical user interface
\dontrun{gui()}
}
