% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsizecomp.R
\name{sampsizecomp}
\alias{sampsizecomp}
\title{Sample Size Estimation for Mixed Outcomes}
\usage{
sampsizecomp(augmean, augvar, binmean, binvar, alpha, beta)
}
\arguments{
\item{augmean}{mean risk difference treatment effect estimated using augmented binary method}

\item{augvar}{variance of risk difference treatment effect estimated using augmented binary method}

\item{binmean}{mean risk difference treatment effect estimated using standard binary method}

\item{binvar}{variance of risk difference treatment effect estimated using standard binary method}

\item{alpha}{one-sided alpha level}

\item{beta}{beta level: 1-desired power}
}
\description{
Available as a Shiny app with documentation at https://github.com/martinamcm/MultSampSize
}
\references{
McMenamin M, Barrett JK, Berglind A, Wason JMS. Sample Size Estimation using a 
Latent Variable Model for Mixed Outcome Co-Primary, Multiple Primary and Composite Endpoints. 
2019. arXiv:1912.05258.

McMenamin M, Grayling MJ, Berglind A, Wason JMS. Increasing power in the 
analysis of responder endpoints in rheumatology: a software tutorial.
medRxiv. 2020. doi: 10.1101/2020.07.28.20163378
}
\author{
Martina McMenamin.
}
