% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsizecomp.R
\name{sampsizecomp}
\alias{sampsizecomp}
\title{Sample Size Estimation for Mixed Outcomes}
\usage{
sampsizecomp(augmean, binmean, augvar, binvar, alpha, beta)
}
\arguments{
\item{augmean}{average treatment effect estimated using augmented binary method}

\item{binmean}{average treatment effect estimated using standard binary method}

\item{augvar}{variance of treatment effect estimated using augmented binary method}

\item{binvar}{variance of treatment effect estimated using standard binary method}

\item{alpha}{one-sided alpha level}

\item{beta}{beta level (1-desired power)}
}
\description{
Determines sample size required in studies using multivariate mixed continuous and discrete endpoints,
when analysed using a latent variable model. The package facilitates sample size estimation for mixed 
outcomes that are combined to form either composite, co-primary and multiple primary endpoints.
}
\details{
Available as a Shiny app with documentation at \href{https://github.com/martinamcm/MultSampSize}{MultSampSize}
}
\references{
McMenamin M, Barrett JK, Berglind A, Wason JMS. Sample Size Estimation using a 
Latent Variable Model for Mixed Outcome Co-Primary, Multiple Primary and Composite Endpoints. 
2019. arXiv:1912.05258.

McMenamin M, Grayling MJ, Berglind A, Wason JMS. Increasing power in the 
analysis of responder endpoints in rheumatology: a software tutorial.
medRxiv. 2020. doi: 10.1101/2020.07.28.20163378
}
\seealso{
\code{\link{augbinrheum}} for obtaining estimates of argument 'augmean', 'augvar', 'binmean', 'binvar' from existing data
}
\author{
Martina McMenamin.
}
